--- src/main/java/ru/javawebinar/topjava/repository/MealRepositoryInterface.java
@@ -1,17 +1,0 @@
-package ru.javawebinar.topjava.repository;
-
-import ru.javawebinar.topjava.model.Meal;
-
-import java.util.Collection;
-import java.util.Collections;
-
-
-public interface MealRepositoryInterface {
-    Meal save(Meal meal);
-
-    boolean deleteById(int id);
-
-    Collection<Meal> getAll();
-
-    Meal getById(int id);
-}
