+++ src/main/webapp/mealCreate.jsp
@@ -12,8 +12,8 @@
 </head>
 <body>
 <a href="index.html">Home</a>
-<hr>           <!-- FIXME typo-->
-<h2>${param.action=='creater'? 'Creater meal' : 'Edit meal'}</h2>
+<hr>
+<h2>${param.action=='create'? 'Create meal' : 'Edit meal'}</h2>
 <jsp:useBean id="meal" type="ru.javawebinar.topjava.model.Meal" scope="request"/>
 <form method="post" action="meals">
     <input type="hidden" name="id" value="${meal.id}">
